+++ deeplabcut/pose_estimation_tensorflow/backbones/efficientnet_model.py
@@ -29,7 +29,7 @@ import six
 from six.moves import xrange  # pylint: disable=redefined-builtin
 import tensorflow as tf
 
-import deeplabcut.pose_estimation_tensorflow.nnet.utils as utils
+import deeplabcut.pose_estimation_tensorflow.nnets.utils as utils
 
 GlobalParams = collections.namedtuple('GlobalParams', [
     'batch_norm_momentum', 'batch_norm_epsilon', 'dropout_rate', 'data_format',
