+++ deeplabcut/pose_estimation_tensorflow/backbones/mobilenet_v2.py
@@ -29,8 +29,8 @@ import tensorflow as tf
 
 # from nets.mobilenet import conv_blocks as ops
 # from nets.mobilenet import mobilenet as lib
-from deeplabcut.pose_estimation_tensorflow.nnet import conv_blocks as ops
-from deeplabcut.pose_estimation_tensorflow.nnet import mobilenet as lib
+from deeplabcut.pose_estimation_tensorflow.nnets import conv_blocks as ops
+from deeplabcut.pose_estimation_tensorflow.backbones import mobilenet as lib
 
 slim = tf.contrib.slim
 op = lib.op
